separate test for custom calendar and mute other unit tests
